Added back RESET_DB, added live link in ReadMe

+++ server.js
@@ -34,6 +34,10 @@ const showSchema = new mongoose.Schema({
 
 const Show = mongoose.model('Show', showSchema);
 
+//TO USE THIS, TYPE: RESET_DB=true npm run dev IN TERMINAL
+
+if (process.env.RESET_DB) {
+  console.log('TippTapp working');
   const seedDB = async () => {
   await Show.deleteMany();
 
@@ -43,7 +47,7 @@ const Show = mongoose.model('Show', showSchema);
   });
 }
 seedDB();
-
+}
 
 const port = process.env.PORT || 8080
 const app = express()
@@ -53,7 +57,6 @@ app.use(express.json())
 
 app.get('/', (req, res) => {
   res.send('Show most go on!')
-  .catch(e => console.log(e))
 })
